// src/redux/types/OverlaysTypes.ts
export type OverlaysType = {
  openConfirmLogoutOverlay: boolean
  openContactOverlay: boolean
  openForwardMessageOverlay: boolean
  openPreviewFilesInMessage: boolean
  openNewGroupOverlay: boolean
  openSettingOverlay: boolean
  openViewGroupInfoOverlay: boolean
  openManageGroupOverlay: boolean
}